add templates for homedir and git-remote content
the CI script will copy the content of template-student-git-remote to the git-remote directory for the student and also copy template-student-homedir to the student home directory for the student. In the staging branch we also delete any content that was present in the student’s gitremvoe and homedir so that we start from a clean slate each time.

Signed-off-by: Mark McCahill <mark.mccahill@duke.edu>